added semantic UI, changed goal to PWA editor, with live repl rebuilds to dockerized server running on host
